+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.1.0
+* The function *registerDbAsset()* accepts the optional parameter *defaultDb*. This parameter allows to set the default database to use.
+* The function *getDatabase()* accepts the optional parameter *key*. If key is omitted the function returns the default database.
 ## 0.0.1
 * Initial release.